% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_freq.R
\name{calc_freq}
\alias{calc_freq}
\title{Calculate Frequency of Binary Variables}
\usage{
calc_freq(df)
}
\arguments{
\item{df}{A dataframe where the first column is the ID and the other variables are binary (0 or 1).}
}
\value{
A dataframe containing the frequency of each binary variable.
}
\description{
The first column of the dataframe is the ID, and the other variables are binary (0 or 1). This function calculates the frequency of each binary variable.
}
