% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_func.R
\name{calc_func}
\alias{calc_func}
\title{Statistical analysis of the efficacy of traditional Chinese herbal medicines}
\usage{
calc_func(df, added = NULL)
}
\arguments{
\item{df}{A data frame, preferably processed by the trans_wide function. That is, the first column is the ID column, and each subsequent column represents a variable.}

\item{added}{A data frame that contains the added information of Chinese medicine properties.}
}
\value{
A data frame that contains the statistics of the efficacy of Chinese medicine.
}
\description{
Statistical analysis of the efficacy of traditional Chinese herbal medicines
}
\examples{
#
}
