% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_herb_dosage.R
\name{calc_herb_dosage}
\alias{calc_herb_dosage}
\title{Statistics of Herb Dosage Distribution}
\usage{
calc_herb_dosage(df, top = NULL)
}
\arguments{
\item{df}{A data frame where the first column is the ID, and the remaining cells contain herb names and their dosages. The dosages should be numeric values representing weight in grams, without any units attached. Correct example: "Huangqi10".}

\item{top}{Only the dosage distribution of the top-ranking herbs is calculated, where the number of herbs considered is determined by the `top` parameter.}
}
\value{
A data frame showing the herb names and their dosage distributions.
}
\description{
Statistics of Herb Dosage Distribution
}
\examples{
df <- data.frame(prescription_id = c(1, 2,3),
herb1 = c("Huangqi10", "Gancao5","Dihuang12"),
herb2 = c("Shengdi", "Danggui12","Banxia15"),
herb3 = c("Baizhu15", "Danshen12", NA),
herb4 = c("Huangqin12", NA,NA ))

}
