% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Phi.R
\name{calc_Phi}
\alias{calc_Phi}
\title{Calculate Phi Coefficient and P-Values}
\usage{
calc_Phi(df, top = NULL)
}
\arguments{
\item{df}{A dataframe containing the binary variables. The first column is ID.}

\item{top}{An integer specifying the number of top columns to consider.
If NULL, the default value is 30. The value must be less than the number of columns in the dataframe.}
}
\value{
A list containing three dataframes:
\item{phi_matrix}{A dataframe with the Phi coefficients.}
\item{p_value}{A dataframe with the p-values.}
\item{star_matrix}{A dataframe with the significance markers ("*") indicating p-values less than 0.05.}
}
\description{
This function calculates the Phi coefficient and the associated p-values
for the columns of a given dataframe. It also generates a significance
matrix indicating which pairs of variables have significant associations.
}
