% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_property.R
\name{calc_property}
\alias{calc_property}
\title{Statistical Information of Chinese Herbal Properties}
\usage{
calc_property(tidywide, added = NULL)
}
\arguments{
\item{tidywide}{A data frame, preferably processed by the trans_wide function. That is, the first column is the ID column, and each subsequent column represents a variable.}

\item{added}{A data frame that contains the added information of Chinese medicine properties.}
}
\value{
A list containing three data frames, representing the statistical results of the frequency of meridian affinity, siqi, and taste respectively.
}
\description{
This function is used to statistically analyze the properties of Chinese herbs, including meridian affinity, aroma, and taste. The user needs to provide a data frame containing the property information of Chinese herbs, including the name of the herb, meridian affinity, aroma, and taste. Additionally, the user should provide another data frame containing the property information of Chinese herbs, including the name of the herb, property information, and frequency of the property. The function will return a list containing three data frames, which are the statistical results of the frequency of meridian affinity, aroma, and taste respectively.
}
\examples{
##
}
