% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPhi.R
\name{calcPhi}
\alias{calcPhi}
\title{Calculate Phi Coefficients Between Variables in Bulk}
\usage{
calcPhi(df, top = NULL)
}
\arguments{
\item{df}{A dataframe where the first column is the ID and the remaining columns are the variables. The ID column is excluded from the calculations.}

\item{top}{A positive integer indicating the number of top variables to calculate the Phi coefficients for. The default value is 30 and it should not exceed the number of variables.}
}
\value{
A list containing the following three elements:
\item{phi_matrix}{Matrix of Phi coefficients}
\item{p_value_matrix}{Matrix of p-values}
\item{star_matrix}{Matrix of significance markers, where elements with p-values less than 0.05 are marked with "*"}
}
\description{
This function calculates the Phi coefficients between multiple variables in a dataframe and returns a matrix of Phi coefficients, a matrix of p-values, and a significance marker matrix. Phi coefficients are used to measure the correlation between two binary variables.
}
\examples{
set.seed(123)
df <- data.frame(ID = 1:10, matrix(sample(0:1, 500, replace = TRUE), nrow = 10, ncol = 50))
result <- calcPhi(df, top = 20)
print(result$phi_matrix)
print(result$p_value_matrix)
print(result$star_matrix)

}
