% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Data Cleaning Function}
\usage{
clean_data(df, lookup_table)
}
\arguments{
\item{df}{A data frame where the first column is the ID, and the remaining columns must contain only one element per cell. For example, "丹参" should occupy one cell, and "丹参,黄芪" in one cell is not allowed.}

\item{lookup_table}{A data frame containing the standardized names of herbs. The user needs to fill in the standard names.}
}
\value{
A data frame with standardized field names.
}
\description{
This function is used to clean and standardize information fields in the data, such as herbal names, symptom names, and acupoint names. The function can identify and handle potential irregular symbols (e.g., newline characters, spaces) and supports fields with or without dosage information.
}
\details{
The main functions of this function include:
1. Removing newline characters and leading/trailing spaces from the data.
2. Checking for duplicate herbs within the same prescription and issuing a warning if found.
3. Determining if the data contains dosage information. If dosage is included, it splits the herbal name and dosage and performs dosage checks (e.g., missing values, non-numeric dosage).
4. Calls the function to standardize herbal names.
5. Reconstructs the data frame based on whether dosage information is included.
}
