% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooc_graph.R
\name{cooc_graph}
\alias{cooc_graph}
\title{Calculate Graph Properties of Co-occurrence Network}
\usage{
cooc_graph(df, top_freq = 30, min = 2, efficacy_add = NULL)
}
\arguments{
\item{df}{A data frame, a tidy wide-format data frame where the first column is the id and the rest are 0-1 binary variables.}

\item{top_freq}{An integer indicating the number of top nodes by frequency to calculate from `df`.}

\item{min}{A positive integer indicating the minimum co-occurrence count.}

\item{efficacy_add}{A data frame containing additional herb efficacy information.}
}
\value{
Returns a list containing two data frames: one for edge attributes and one for node attributes.
}
\description{
This function calculates the properties of herbs in a co-occurrence network, including node degree and herb efficacy.
}
\examples{
# Example usage
#cooc_graph(data, top_freq = 30, min = 2)
}
