% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{data_check}
\alias{data_check}
\title{Data Check}
\usage{
data_check(df, format)
}
\arguments{
\item{df}{A data frame, formatted as long data or wide data.}

\item{format}{A string, taking the value of "long" or "basket".}
}
\value{
Returns a list containing two data frames, \code{data_clean} is the data frame after removing duplicate herbs, and \code{duplicate_herbs} is the data frame containing duplicate herbs.
}
\description{
The function is used to check for duplicates of Traditional Chinese Medicine (TCM), symptoms, and spaces in the data.
}
\examples{
df <- data.frame(id = c(1, 1, 2, 3, 3), herb = c('A', 'A', 'B', 'C', 'C'))
df2 <- data.frame(id = c(1, 1, 2, 3, 3), herb = c('A B', 'A ', 'B', 'C', 'C'))
data_check(df2, "basket")
}
