% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_herbs.R
\name{distinct_herbs}
\alias{distinct_herbs}
\title{Remove duplicate elements from a formula}
\usage{
distinct_herbs(x)
}
\arguments{
\item{x}{A data frame where the first column is the prescription ID, each row is a prescription, and each cell contains one Chinese herb; the data is disorganized.}
}
\value{
A data frame where the herbs in each prescription are unique.
}
\description{
For manually collected data, a common format is: each row represents one prescription, and each cell contains one Chinese herb or symptom, leading to disorganized data. This function is used to process such data.
Due to human error or system issues, there may be duplicate herbs or symptoms in the data. This function removes duplicate herbs in each row, maintaining the uniqueness of herbs in each prescription.
}
\examples{
df <- data.frame(
  id = 1:3,
  herb1 = c("A", "B", "A"),
  herb2 = c("A", "C", "B"),
  herb3 = c("B", NA, "C"),
  herb4 = c(NA, NA, "A"),
  stringsAsFactors = FALSE
)
distinct_herbs(df)
}
