% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_cluster.R
\name{explore_cluster}
\alias{explore_cluster}
\title{Preliminary Exploration of Various Clustering Structures}
\usage{
explore_cluster(df, top = 30)
}
\arguments{
\item{df}{A data frame where rows represent formulas (prescriptions), columns represent Chinese herbs, and the first column is the formula number.}

\item{top}{A positive integer indicating the number of top Chinese herbs to be included in the cluster analysis. The default is 30, but it is recommended to adjust this according to specific needs.}
}
\value{
The function does not return any value; instead, it directly plots dendrograms for the seven clustering methods.
}
\description{
Different clustering methods can lead to varying results. This function aims to quickly plot clustering structure diagrams using seven different clustering methods. These include: "ward.D", "ward.D2", "single", "complete", "average", "centroid", and "median".
}
\examples{
#' # Create a sample data frame with 10 rows (formulas) and 31 columns (1 ID column + 30 herbs)
set.seed(123)
df <- data.frame(ID = 1:10, matrix(rnorm(300), nrow = 10, ncol = 30))

# Explore clustering structure with default top 30 herbs
explore_cluster(df)

# Explore clustering structure with top 20 herbs
explore_cluster(df, top = 20)
}
