% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_rules.R
\name{explore_rules}
\alias{explore_rules}
\title{Exploring Association Rules}
\usage{
explore_rules(
  trans,
  supp = c(0.05, 0.07, 0.1, 0.12, 0.15, 0.2, 0.22, 0.25, 0.28, 0.3),
  conf = c(0.5, 0.55, 0.6, 0.65, 0.7, 0.75, 0.8, 0.85, 0.9),
  maxlen = 4
)
}
\arguments{
\item{trans}{The transformed transactions object.}

\item{supp}{Set of support values, defaulting to c(0.05, 0.07, 0.10, 0.12, 0.15, 0.2, 0.22, 0.25, 0.28, 0.3); customization is recommended.}

\item{conf}{Set of confidence values, defaulting to c(0.5, 0.55, 0.6, 0.65, 0.7, 0.75, 0.8, 0.85, 0.9); customization is recommended.}

\item{maxlen}{The maximum length of association rules, default is 4. For clinical data, it is recommended not to exceed 4.}
}
\value{
A data frame containing the number of association rules, the set support level, the set confidence level, the minimum support, the maximum support, the minimum confidence, the maximum confidence, and the minimum lift. Rows that resulted in Inf due to excessively high or low parameters have been removed.
}
\description{
For different combinations of support and confidence, this function returns the number of association rules, support, confidence, lift, and other information. Due to the specific nature of clinical data, it is recommended to customize the support and confidence combinations. Additionally, only association rules with a lift greater than 1 are returned, as only these are meaningful.
}
