% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_cooc.R
\name{fig_cooc}
\alias{fig_cooc}
\title{Draw a Co-occurrence Network Graph of Chinese Herbal Medicine}
\usage{
fig_cooc(
  g,
  col_pal = "Set3",
  line_type = "arc",
  font_color = "black",
  edge_col_low = "steelblue",
  edge_col_high = "tomato",
  edge_alpha_range = c(0.2, 1),
  node_layout = "fr",
  node_alpha = 0.55,
  text_size = 3,
  edge_width_range = c(0.3, 1),
  node_size = c(4, 10)
)
}
\arguments{
\item{g}{A list containing node and edge data returned by the function `cooc_graph`.}

\item{col_pal}{The color palette name for network nodes. Default is "Pastel1".}

\item{line_type}{The type of line to use for edges, can be "arc" or "straight". Default is "arc".}

\item{font_color}{The color of the node labels. Default is "black".}

\item{edge_col_low}{The color of connections with low co-occurrence frequency. Default is "steelblue".}

\item{edge_col_high}{The color of connections with high co-occurrence frequency. Default is "tomato".}

\item{edge_alpha_range}{The range of edge transparency. Default is c(0.2, 1).}

\item{node_layout}{The layout for the nodes. Default is "fr". Options include "circle", "kk", "fr", "gem", "randomly", "nicely", "stress", "grid", "dh", "star", "graphopt", "sugiyama", and "sphere".}

\item{node_alpha}{The transparency level of the nodes. Default is 0.55.}

\item{text_size}{The size of the node labels. Default is 3.}

\item{edge_width_range}{The range of edge widths. Default is c(0.3, 1).}

\item{node_size}{The range of node sizes. Default is c(4, 10).}
}
\value{
Returns a co-occurrence network graph.
}
\description{
Draw a Co-occurrence Network Graph of Chinese Herbal Medicine
}
