% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_dose.R
\name{fig_dose}
\alias{fig_dose}
\title{Plot the Distribution of Herb Dosages}
\usage{
fig_dose(
  df,
  top = 25,
  percent_ingore = 5,
  bar_line_size = 0.5,
  bar_alpha = 0.7,
  text_size = 1.5,
  bar_pal = "Set1",
  min_col = "steelblue",
  max_col = "salmon",
  mean_col = "darkred",
  dumbbell_col = "gray60"
)
}
\arguments{
\item{df}{A data frame where the first column is the ID, and the remaining columns are the herb names and their corresponding dosage values. The dosage values are in grams, without any attached units. Each cell occupies one value.}

\item{top}{The number of top herbs to include in the frequency ranking. The default is 25.}

\item{percent_ingore}{The threshold for excluding dosages with a percentage below this value in the dosage percentage bar chart. The default is 5.}

\item{bar_line_size}{The width of the border in the bar chart.}

\item{bar_alpha}{The transparency of the bars in the bar chart.}

\item{text_size}{The font size of the dosage labels.}

\item{bar_pal}{The color palette for the bar chart.}

\item{min_col}{The color for the minimum dosage in the dumbbell plot.}

\item{max_col}{The color for the maximum dosage in the dumbbell plot.}

\item{mean_col}{The color for the average dosage in the dumbbell plot.}

\item{dumbbell_col}{The color of the lines in the dumbbell plot.}
}
\value{
A ggplot2 object representing a combined plot of the dosage frequency percentage chart and the dumbbell plot.
}
\description{
This function plots the distribution of herb dosages. The final result is a combined plot of a dosage frequency percentage bar chart and a dumbbell plot.
}
