% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_efficacy.R
\name{fig_efficacy}
\alias{fig_efficacy}
\title{Create a Sunburst Plot for Efficacy Data}
\usage{
fig_efficacy(df, cal_pal = "Set3", ignore = 0.03, lab_size = 2.5, added = NULL)
}
\arguments{
\item{df}{A data frame containing the input data. The data frame should have columns named `main_efficacy`, `sub_efficacy`, and `n`.}

\item{cal_pal}{A string representing the RColorBrewer palette to be used. Default is "Set3".}

\item{ignore}{A numeric value between 0 and 1 indicating the threshold below which categories will be ignored. Default is 0.03.}

\item{lab_size}{A numeric value specifying the label text size in the plot. Default is 2.5.}

\item{added}{A data frame that contains the added information of Chinese medicine properties.}
}
\value{
A ggplot2 object representing the sunburst plot.
}
\description{
This function calculates the efficacy of the given data and creates a sunburst plot
to visualize the main and sub efficacy categories. It allows customization of the
plot palette, label size, and an option to ignore categories below a certain threshold.
}
\examples{
\dontrun{
  df <- data.frame(
    main_efficacy = c("A", "B", "A", "C"),
    sub_efficacy = c("A1", "B1", "A2", "C1"),
    n = c(10, 20, 30, 40)
  )
  fig_efficacy(df, cal_pal = "Set2", ignore = 0.05, lab_size = 3)
}
}
