% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_exp_rule.R
\name{fig_exp_rule}
\alias{fig_exp_rule}
\title{Explore Association Rules with Bubble Chart}
\usage{
fig_exp_rule(rules, col_pal = "D")
}
\arguments{
\item{rules}{A data frame containing the association rules and their metrics.}

\item{col_pal}{The color palette option for the viridis color scale. Default is "D".}
}
\value{
A ggplot2 object representing the bubble chart.
}
\description{
This function creates a bubble chart to explore association rules.
}
\examples{
#

}
