% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_font.R
\name{fig_font}
\alias{fig_font}
\title{Set the Font for Plotting}
\usage{
fig_font(font = "Song")
}
\arguments{
\item{font}{The font name, either "Songti" or "Kaiti". Default is "Songti", used for plotting on Linux.}
}
\value{
No return value.
}
\description{
Set the Font for Plotting
}
\examples{
# fig_font("Songti")

}
