% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_freq_bar.R
\name{fig_freq_bar}
\alias{fig_freq_bar}
\title{Draws a bar chart based on frequency.}
\usage{
fig_freq_bar(
  df,
  top = 10,
  horizontal = TRUE,
  low = "lightsteelblue2",
  high = "steelblue",
  bar_width = 0.7
)
}
\arguments{
\item{df}{A tidy data frame with an ID column as the first column.}

\item{top}{The number of top items by frequency to select. Default is 10.}

\item{horizontal}{Whether to display bars horizontally. Default is TRUE.}

\item{low}{The color for the lowest frequencies. Default is "lightsteelblue2".}

\item{high}{The color for the highest frequencies. Default is "steelblue".}

\item{bar_width}{The width of the bars. Default is 0.7.}
}
\value{
A ggplot2 object.
}
\description{
Draws a bar chart based on frequency.
}
\examples{
# Example usage:



}
