% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_jac_shiny.R
\name{fig_jac_shiny}
\alias{fig_jac_shiny}
\title{Visualize of Traditional Chinese Medicine Prescriptions similarity in Shiny}
\usage{
fig_jac_shiny(jac_res, col_pal = "RdYlBu", cell_size = 2)
}
\arguments{
\item{jac_res}{A list of Jaccard similarity matrix and Jaccard index, the result of \code{\link{calc_jaccard}}}

\item{col_pal}{Color palette from RColorBrewer for the heatmap. Default is "RdYlBu".}

\item{cell_size}{Size of the cells in the heatmap. Default is 2.}
}
\value{
A heatmap of Jaccard similarity matrix.
}
\description{
Visualize of Traditional Chinese Medicine Prescriptions similarity in Shiny
}
