% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_matrix.R
\name{fig_matrix}
\alias{fig_matrix}
\title{Generate Heatmap with Clustering and Optional Significance Markers}
\usage{
fig_matrix(matrix, star_matrix = NULL, fontsize = 8, cellsize = 8, k = 5)
}
\arguments{
\item{matrix}{A numeric matrix to be visualized as a heatmap.}

\item{star_matrix}{An optional matrix of the same dimensions as \code{matrix}
containing significance markers to be displayed on the heatmap.
If \code{NULL}, no markers will be displayed.}

\item{fontsize}{An integer specifying the font size for row and column labels.}

\item{cellsize}{An integer specifying the width and height of the heatmap cells.}

\item{k}{An integer specifying the number of clusters to cut the dendrogram into.}
}
\value{
A pheatmap object representing the heatmap with clustering and optional significance markers.
}
\description{
This function generates a heatmap with clustering for rows and columns.
Optionally, it can display significance markers on the heatmap.
}
