% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_phi_shiny.R
\name{fig_phi_shiny}
\alias{fig_phi_shiny}
\title{Plot Clustering and Heatmap of Phi Coefficients in Shiny}
\usage{
fig_phi_shiny(
  phi_res,
  col_pal = "RdBu",
  k = 5,
  fontsize = 4,
  border_color = "#8B0A50",
  cell_cize = 8,
  font_angle = 270,
  tree_height = 15
)
}
\arguments{
\item{phi_res}{A list containing 3 dataframes,the result of \code{\link{calc_Phi}} function.}

\item{col_pal}{Color palette of the RColorBrewer for the heatmap. Default is "RdBu".}

\item{k}{The number of clusters for both rows and columns. Default is 5.}

\item{fontsize}{Font size for the labels in the heatmap. Default is 8.}

\item{border_color}{Border color for the cells in the heatmap. Default is "#8B0A50".}

\item{cell_cize}{The size of each cell in the heatmap. Default is 8.}

\item{font_angle}{The angle of the column labels. Default is 270 degrees.}

\item{tree_height}{The height of the dendrogram for the rows. Default is 15.}
}
\value{
A heatmap of the Phi coefficients with clustering.
}
\description{
This function calculates the Phi coefficients for binary variables in a data frame,
visualizes the results in a heatmap.
}
