% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_rules_bubble.R
\name{fig_rules_bubble}
\alias{fig_rules_bubble}
\title{Generate Bubble Plot for Association Rules}
\usage{
fig_rules_bubble(
  rules,
  shading = "lift",
  measure = "support",
  col = c("#EEA2AD", "#8DB6CD"),
  alpha = 0.7,
  k = 20,
  top = NULL
)
}
\arguments{
\item{rules}{An object of class `rules` containing the association rules.}

\item{shading}{A character string specifying the shading measure. Default is `"lift"`.}

\item{measure}{A character string specifying the measure. Default is `"support"`.}

\item{col}{A vector of colors for the bubbles. Default is `c("#EEA2AD", "#8DB6CD")`.}

\item{alpha}{Transparency level for the bubbles. Default is `0.7`.}

\item{k}{Number of items to show in the grouped plot. Default is `20`.}

\item{top}{A numeric value between `0` and `1` representing the top percentage of rules to display. Default is `NULL`, which means all rules are displayed.}
}
\value{
A `ggplot` object representing the bubble plot of the association rules.
}
\description{
This function generates a bubble plot for association rules using the `grouped` method from the `arulesViz` package and `ggplot2` for visualization.
}
\examples{
# Assuming `rules` is an object of class `rules`
# fig_rules_bubble(rules)
}
