% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_rules_distr.R
\name{fig_rules_distr}
\alias{fig_rules_distr}
\title{Visualize Association Rules Distribution}
\usage{
fig_rules_distr(rules, col_pal = "D", top = NULL)
}
\arguments{
\item{rules}{A set of association rules.}

\item{col_pal}{The color palette option for the viridis color scale. Default is "D".}

\item{top}{A positive number less than 1 to specify the top percentage of rules to display. Default is NULL, which displays all rules.}
}
\value{
A ggplot2 object representing the scatter plot of the association rules.
}
\description{
This function creates a scatter plot to visualize the distribution of association rules. Accept an association rules object obtained from the extra_rules function.
}
\examples{
# Example usage:
# rules <- apriori(data = transactions, parameter = list(supp = 0.1, conf = 0.8))
# p <- fig_rules_distr(rules, top = 0.1)
# print(p)
}
