% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_rules_network.R
\name{fig_rules_network}
\alias{fig_rules_network}
\title{Visualize Association Rules as a Network Graph}
\usage{
fig_rules_network(
  rules,
  rule.size = "support",
  rule.color = "confidence",
  min.val = "#FFDEAD",
  max.val = "blue",
  edge_col = "blue",
  num_rules = 10
)
}
\arguments{
\item{rules}{An object of class "rules" generated by the `apriori` function from the `arules` package.}

\item{rule.size}{A character string specifying the measure to use for node size. Options are "support", "confidence", or "lift".}

\item{rule.color}{A character string specifying the measure to use for node color. Options are "support", "confidence", or "lift".}

\item{min.val}{A character string specifying the color for the minimum value in the color gradient.}

\item{max.val}{A character string specifying the color for the maximum value in the color gradient.}

\item{edge_col}{A character string specifying the color of the edges.}

\item{num_rules}{An integer specifying the maximum number of rules to visualize. Default is 10.}
}
\value{
A ggplot2 object representing the network graph of association rules.
}
\description{
This function visualizes a set of association rules as a network graph using
the `ggraph` package. Nodes represent items and edges represent rules between items.
}
