% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sankey_shiny.R
\name{fig_sankey_shiny}
\alias{fig_sankey_shiny}
\title{Draw a comprehensive Sankey diagram.}
\usage{
fig_sankey_shiny(
  dim1,
  dim2,
  dim3 = NULL,
  dim4 = NULL,
  dim1_n = 20,
  dim2_n = 5,
  dim3_n = 20,
  dim4_n = 20,
  col_pal = "A",
  label_size = 3.1,
  label_color = "black",
  node_alpha = 0.7
)
}
\arguments{
\item{dim1}{A data frame containing the data for the first dimension}

\item{dim2}{A data frame containing the data for the second dimension}

\item{dim3}{A data frame containing the data for the third dimension}

\item{dim4}{A data frame containing the data for the fourth dimension}

\item{dim1_n}{An integer specifying the top 'n' elements by frequency for the first dimension}

\item{dim2_n}{An integer specifying the top 'n' elements by frequency for the second dimension}

\item{dim3_n}{An integer specifying the top 'n' elements by frequency for the third dimension}

\item{dim4_n}{An integer specifying the top 'n' elements by frequency for the fourth dimension}

\item{col_pal}{A character string specifying the color palette to use，one of "A", "B", "C", "D", "E", "F", "G", "H"}

\item{label_size}{A numeric specifying the font size of the node labels}

\item{label_color}{A character string specifying the color of the node labels}

\item{node_alpha}{A numeric specifying the transparency of the nodes}
}
\value{
A ggplot2 object
}
\description{
Draw a comprehensive Sankey diagram.
}
