% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_siqi.R
\name{fig_siqi}
\alias{fig_siqi}
\title{Generate a Radar Plot for Siqi Properties}
\usage{
fig_siqi(
  df,
  line_col = "red",
  bac_col = "skyblue",
  label.size = 5,
  line.width = 1,
  added = NULL
)
}
\arguments{
\item{df}{A data frame containing the siqi data with at least two columns: 'siqi' and 'weighted_frequency'.}

\item{line_col}{Character. Color for the lines connecting the data points on the radar plot. Default is 'red'.}

\item{bac_col}{Character. Background color for the radar plot. Default is 'skyblue'.}

\item{label.size}{Numeric. Size of the labels on the radar plot. Default is 5.}

\item{line.width}{Numeric. Width of the lines connecting the data points on the radar plot. Default is 1.}

\item{added}{A data frame that contains the added information of Chinese medicine properties.}
}
\value{
A radar plot visualizing the siqi properties.
}
\description{
This function creates a radar plot to visualize the weighted frequency of various siqi (four natures) in a given data frame.
}
