% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpSimScore.R
\name{grpSimScore}
\alias{grpSimScore}
\title{Calculate the similarity between two groups of prescriptions}
\usage{
grpSimScore(df1, df2)
}
\arguments{
\item{df1}{a data frame A wide data frame, the first column being the ID.}

\item{df2}{a data frame A wide data frame, the first column being the ID.}
}
\value{
a numeric value The similarity between the two groups of prescriptions.
}
\description{
Calculate the similarity between two groups of prescriptions. Input two wide data frames, the first column being the ID. The column names of the ID columns in both data frames should be unified to "id".
}
