% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_standard_names.R
\docType{data}
\name{herb_standard_names}
\alias{herb_standard_names}
\title{Dataset of Classification of Chinese Herbal Medicine Efficacy}
\format{
A data frame with 8287 rows and 1 variables:
\describe{
\item{names}{Chinese Herbal Medicine Name}
...
}
}
\usage{
herb_standard_names
}
\description{
Dataset of Classification of Chinese Herbal Medicine Efficacy
}
\keyword{datasets}
