% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herbs_attr.R
\docType{data}
\name{herbs_attr}
\alias{herbs_attr}
\title{Chinese Medicine Property Dataset}
\format{
A data frame with 639 rows and 4 variables:
\describe{
\item{names}{Chinese Herbal Medicine Name}
\item{taste}{the taste of Chinese herbs}
\item{siqi}{The Four Properties of Chinese Herbs}
\item{meridian}{Meridian Attribution of Traditional Chinese Medicine}
...
}
}
\usage{
herbs_attr
}
\description{
A short description...

This dataset contains information about the Four Properties and Five Tastes, as well as the meridian tropism of traditional Chinese medicines..
}
\keyword{datasets}
