% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herbs_function.R
\docType{data}
\name{herbs_function}
\alias{herbs_function}
\title{Dataset of Classification of Chinese Herbal Medicine Efficacy}
\format{
A data frame with 905 rows and 3 variables:
\describe{
\item{names}{Chinese Herbal Medicine Name}
\item{main_function}{First-level Classification of Chinese Herbal Medicine Efficacy}
\item{sub_function}{TSecondary Classification of Chinese Herbal Medicine Efficacy}
...
}
}
\usage{
herbs_function
}
\description{
Dataset of Classification of Chinese Herbal Medicine Efficacy
}
\keyword{datasets}
