% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insomnia}
\alias{insomnia}
\title{Insomnia Dataset}
\format{
A non-tidy data frame with 497 rows:
\describe{
  \item{patient_id}{Unique patient identifier}
}
}
\source{
Traditional Chinese Medicine outpatient data
}
\usage{
insomnia
}
\description{
This is an ancient Chinese prescription dataset for treating insomnia, including the dosage of Chinese herbs.
}
\keyword{datasets}
