% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_selection.R
\name{k_selection}
\alias{k_selection}
\title{k_selection Function}
\usage{
k_selection(df, top = NULL, method = NULL)
}
\arguments{
\item{df}{A dataframe where the first column is assumed to be an identifier and will be removed before clustering.}

\item{top}{An positive integer specifying the number of top columns to use for clustering. Defaults to 30.}

\item{method}{A string specifying the clustering method to use. Defaults to "ward.D2".}
}
\value{
A list and four scatter-line plots. The list contains two data frames, resdf and k_vote. resdf includes the detailed results of the clustering votes, while k_vote summarizes the distribution of recommended votes. The scatter-line plots are provided to aid in the recommendation of the number of clusters.
}
\description{
This function performs clustering on the input dataframe using the NbClust package to determine the optimal number of clusters. It supports various clustering indices to provide a comprehensive evaluation.
}
\examples{
# Example usage:
# Assuming df3 is your input dataframe
# df5 <- k_selection(df3)
}
