% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidney}
\alias{kidney}
\title{Chinese Patent Medicine Dataset}
\format{
A non-tidy data frame with 259 rows:
\describe{
  \item{id}{Unique identifier}
}
}
\source{
\emph{Chinese Pharmacopoeia}
}
\usage{
kidney
}
\description{
This is a dataset of Chinese patent medicines used to treat kidney deficiency syndrome, where each row represents a formula. Each cell represents a Chinese herb without including the dosage.
}
\keyword{datasets}
