% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makelookuptable.R
\name{makelookuptable}
\alias{makelookuptable}
\title{Extract Herb Names to be Standardized}
\usage{
makelookuptable(df, type = NULL)
}
\arguments{
\item{df}{A dataframe where the first column is the ID column, and the subsequent columns are herb names in any format.}

\item{type}{A string specifying the type of normalization. Currently supports "Chinese" and "letters". "Chinese" is for herb names in Chinese characters, and during processing, all spaces in the dataframe will be removed. "letters" is for herb names in alphabetic characters, such as Pinyin or scientific names in Latin. During processing, leading and trailing spaces will be removed, and multiple consecutive spaces will be replaced with a single space.}
}
\value{
Returns a dataframe containing the herb names to be standardized, their frequencies, and a column for you to fill in with the standardized herb names.
}
\description{
Extract Herb Names to be Standardized
}
