% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makelookuptable2.R
\name{makelookuptable2}
\alias{makelookuptable2}
\title{Extract Names to be Standardized}
\usage{
makelookuptable2(df, type = NULL)
}
\arguments{
\item{df}{A data frame containing names that may need to be standardized.}

\item{type}{A string specifying the type of normalization. Currently supports "Chinese" and "letters". "Chinese" is for herb names in Chinese characters, and during processing, all spaces in the dataframe will be removed. "letters" is for herb names in alphabetic characters, such as Pinyin or scientific names in Latin. During processing, leading and trailing spaces will be removed, and multiple consecutive spaces will be replaced with a single space.}
}
\value{
Returns a data frame with three columns: the names to be standardized, their corresponding frequency, and a column for filling in the standardized names.
}
\description{
Used to identify names that need to be standardized. `makelookuptable2` relies on the `makelookuptable` function. If there are numeric values for herbal dosages, the function will automatically detect and remove the numbers, then generate the standardized names. Each element occupies one cell.
}
