% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miner_shiny.R
\name{miner_shiny}
\alias{miner_shiny}
\title{Launch the Sinomedminer Shiny Application}
\usage{
miner_shiny(port = NULL)
}
\arguments{
\item{port}{The port to run the Shiny application on.}
}
\description{
This function launches the Sinomedminer Shiny application.
}
