% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sanky_screen.R
\name{tidy_sanky_screen}
\alias{tidy_sanky_screen}
\title{Convert dimension data to long data format suitable for creating Sankey diagrams}
\usage{
tidy_sanky_screen(
  dim1,
  dim2,
  dim3 = NULL,
  dim4 = NULL,
  dim1_n = 20,
  dim2_n = 5,
  dim3_n = 20,
  dim4_n = 20
)
}
\arguments{
\item{dim1}{A data frame containing the data for the first dimension}

\item{dim2}{A data frame containing the data for the second dimension}

\item{dim3}{A data frame containing the data for the third dimension}

\item{dim4}{A data frame containing the data for the fourth dimension}

\item{dim1_n}{An integer specifying the top 'n' elements by frequency for the first dimension}

\item{dim2_n}{An integer specifying the top 'n' elements by frequency for the second dimension}

\item{dim3_n}{An integer specifying the top 'n' elements by frequency for the third dimension}

\item{dim4_n}{An integer specifying the top 'n' elements by frequency for the fourth dimension}
}
\value{
A list containing the long data frame and frequency data frame
}
\description{
Convert dimension data to long data format suitable for creating Sankey diagrams
}
