% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_long2.R
\name{trans_long2}
\alias{trans_long2}
\title{Convert data to long format}
\usage{
trans_long2(df)
}
\arguments{
\item{df}{A messy data frame where the first column is the prescription ID, each row is a prescription, and each cell contains one Chinese herb. If a herb has an associated dose, the dose will be extracted and converted into a numeric value representing the weight in grams, without including the unit (e.g., "Huangqi10").}
}
\value{
A long-format data frame containing three columns: prescription ID, herb, and the herb's dose (if available). Each row represents one herb.
}
\description{
Convert data to long format
}
\examples{
df <- data.frame(prescription_id = c(1, 2,3),
herb1 = c("Huangqi10", "Gancao5","Dihuang12"),
herb2 = c("Shengdi", "Danggui12","Banxia15"),
herb3 = c("Baizhu15", "Danshen12", NA),
herb4 = c("Huangqin12", NA,NA )
)
}
