% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_wide.R
\name{trans_wide}
\alias{trans_wide}
\title{Convert data to wide format}
\usage{
trans_wide(x)
}
\arguments{
\item{x}{A data frame where the first column is the prescription ID, each row is a prescription, and each cell contains one Chinese herb.}
}
\value{
A wide-format data frame where the first column is the prescription ID, followed by herb columns. Each row represents a prescription. Herbs are arranged in descending order of frequency from left to right. A 0 indicates the herb is not included in the prescription, and a 1 indicates it is included.
}
\description{
For manually collected data, a common format is: each row represents one prescription, and each cell contains one Chinese herb, leading to disorganized data.
Occasionally, it's necessary to transform the data into a wide format for subsequent analysis. Since data mining often focuses on high-frequency herbs, this function has already sorted the herbs.
}
\examples{
df <- data.frame(
 id = 1:3,
 herb1 = c("huangqi", "baizhu", "huangqi"),
 herb2 = c("huangqi", "renshen", "baizhu"),
 herb3 = c("baizhu", NA, "renshen"),
 herb4 = c(NA, NA, "huangqi"),
 stringsAsFactors = FALSE)
trans_wide(df)
}
