% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untidy_freq.R
\name{untidy_freq}
\alias{untidy_freq}
\title{Calculating the frequency of uncleaned/messy data.}
\usage{
untidy_freq(df)
}
\arguments{
\item{df}{A data frame with at least two columns. The first column is preserved, while numeric characters
in other columns will be removed.}
}
\value{
A data frame with two columns: one for the unique values and one for the corresponding frequency counts.
The data frame is sorted in descending order by the frequency count.
}
\description{
This function cleans all columns in the data frame except the first one by removing numeric characters,
and then calculates the frequency of unique values. The result is a frequency table sorted in descending
order of the frequency count.
}
\examples{
df <- data.frame(
  group = c("A", "B", "A", "B", "C"),
  value1 = c("1a", "2b", "1c", "2a", "3b"),
  value2 = c("a1", "b2", "c3", "a2", "b1")
)
untidy_freq(df)


}
