% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_similarity.R
\name{wt_similarity}
\alias{wt_similarity}
\title{Calculate weighted similarity between a core formula and other formulas}
\usage{
wt_similarity(df, core_id, a, b, c, d)
}
\arguments{
\item{df}{Data frame, with the first column being an ID column, and the remaining columns being variable names. Each column's values are either 0 or 1.}

\item{core_id}{String, the ID of the core formula.}

\item{a}{Vector of strings, the names of the primary herbs. It is required that the user designate the names of the Chinese medicinal herbs, with the stipulation that these names must be present within the composition of the core formula.}

\item{b}{Vector of strings, the names of the secondary herbs.It is required that the user designate the names of the Chinese medicinal herbs, with the stipulation that these names must be present within the composition of the core formula.}

\item{c}{Vector of strings, the names of the auxiliary herbs.It is required that the user designate the names of the Chinese medicinal herbs, with the stipulation that these names must be present within the composition of the core formula.}

\item{d}{Vector of strings, the names of the messenger herbs.It is required that the user designate the names of the Chinese medicinal herbs, with the stipulation that these names must be present within the composition of the core formula.}
}
\value{
Returns a data frame, containing the IDs of the non-core formulas and their weighted similarity to the core formula.
}
\description{
Calculate weighted similarity between a core formula and other formulas
}
\examples{
# Examples will be here
}
